+++ lib/production/entities/work_order.rb
@@ -4,8 +4,8 @@ module ProductionApp
   class WorkOrder < Dry::Struct
     attribute :id, Types::Integer
     attribute :marketing_order_id, Types::Integer
-    attribute :carton_qty_required, Types::Decimal
-    attribute :carton_qty_produced, Types::Decimal
+    attribute :carton_qty_required, Types::Integer
+    attribute :carton_qty_produced, Types::Integer
     attribute :start_date, Types::Date
     attribute :end_date, Types::Date
     attribute :completed, Types::Bool
