+++ lib/production/entities/work_order_item.rb
@@ -5,8 +5,8 @@ module ProductionApp
     attribute :id, Types::Integer
     attribute :work_order_id, Types::Integer
     attribute :product_setup_id, Types::Integer
-    attribute :carton_qty_required, Types::Decimal
-    attribute :carton_qty_produced, Types::Decimal
+    attribute :carton_qty_required, Types::Integer
+    attribute :carton_qty_produced, Types::Integer
     attribute :completed, Types::Bool
     attribute :completed_at, Types::DateTime
   end
